requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asyncpg==0.30.0
black==24.10.0
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.9
distlib==0.3.9
distro==1.9.0
fastapi==0.115.6
filelock==3.16.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
identify==2.6.3
idna==3.10
iniconfig==2.0.0
isort==5.13.2
jiter==0.8.2
joblib==1.4.2
loguru==0.7.3
mypy-extensions==1.0.0
nltk==3.9.1
nodeenv==1.9.1
numpy==2.2.0
openai==1.58.1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pgvector==0.3.6
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
virtualenv==20.28.0
